proper copy onto skeleton
