--- src/error_enum/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod error_enum;
