--- src/jwt_encoder/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod jwt_encoder;
