+++ src/main.rs
@@ -6,10 +6,10 @@ pub(crate) mod config;
 pub(crate) mod db;
 pub(crate) mod endpoint;
 
+use crate::common::jwt_laporan;
 use crate::config::AppConfig;
 use crate::endpoint::route_handler::init as init_handler;
 use actix_web::{middleware, App, HttpServer};
-use crate::common::jwt_laporan;
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
@@ -21,4 +21,3 @@ async fn main() -> std::io::Result<()> {
         .run()
         .await
 }
-
