--- src/redis_helper/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod redis_helper;
