--- src/redis_helper/redis_helper.rs
@@ -1,3 +1,0 @@
-pub(crate) fn validate_jwt_token(the_token:&str) -> bool {
-    true
-}
